+++ src/Accounts/Accounts/ChangeLog.md
@@ -22,6 +22,7 @@
 * Enable WAM as default way of interactive login flow.
 * Upgraded Azure.Identity to 1.11.2 and Azure.Identity.Broker to 1.1.0.
 * Upgraded Azure.Core to 1.38.0 and Microsoft.Identity.Client to 4.60.3.
+* Enabled secrets detection option by default.
 * Fixed a null reference issue during the process of `Get-AzContext -ListAvailable` [#24854].
 * Supported interactive subscription selection for user login flow. See more details at [Announcing a new login experience with Azure PowerShell and Azure CLI
 ](https://techcommunity.microsoft.com/t5/azure-tools-blog/announcing-a-new-login-experience-with-azure-powershell-and/ba-p/4109357)