+++ src/CustomLocation/CustomLocation/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Upgraded managed identity parameters.
+* Updated example for new managed identity.
 
 ## Version 0.1.3
 * Introduced secrets detection feature to safeguard sensitive data.